backend/requirements.txt
pymongo
python-dotenv
fastapi
uvicorn
pyjwt
jose
pydantic
typing
starlette
httpx
traceback2